+++ packages/rfw/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.24
+* Adds `InkResponse` material widget.
+* Adds `Material` material widget.
+* Adds the `child` to `Opacity` core widget.
+* Implements more `InkWell` parameters.
+
 ## 1.0.23
 
 * Replaces usage of deprecated Flutter APIs.